Make ConnectionFactory abstract
